pyproject.toml
[project]
name = "stock-prediction"
version = "0.1.0"
description = "Stock price prediction using news sentiment and financial indicators"
authors = [
    {name = "Stock Prediction Team"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "yfinance>=0.2.18",
    "newsapi-python>=0.2.6",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    "dask[complete]>=2023.5.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "streamlit>=1.25.0",
    "plotly>=5.15.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-bigquery>=3.11.0",
    "optuna>=3.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pyarrow>=12.0.0",
    "seaborn>=0.12.0",
    "matplotlib>=3.7.0",
    "jupyter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true